refactored mainactivty, marks elements to be transfer into other place finished share into app
